pyproject.toml
[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[project]
name = "release-mcp"
version = "0.1.0"
description = "MCP server for Release Service assistance"
requires-python = ">=3.12"
dependencies = [
    "mcp==1.26.0",
    "pyyaml==6.0.3",
    "jsonschema==4.26.0",
]

[project.scripts]
release-mcp = "release_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/release_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[project.optional-dependencies]
dev = [
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "ruff==0.15.5",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
